Implemented Normal Distribuition

+++ src/entidades/distribuicoes/abstracts/Distribuicao.java
@@ -1,5 +1,5 @@
 package entidades.distribuicoes.abstracts;
 
 public interface Distribuicao {
-	public abstract Double calcular(Double x); //método que calcula a função da distribuição
+	public abstract Double calcular(Integer x); //método que calcula a função da distribuição
 }
